Make commDiags test pass with comm layers supporting network atomics.
The commDiags test would fail sporadically in configurations supporting
network atomics (recent runs: 2/1000 failures with comm=ofi on our Linux
IB-based cluster and 12/100 failures with comm=ugni on a Cray XC).  In
the first block of the output, variability in which task reaches the
barrier last causes variability in the comm diags output.  Here, adjust
for that variability.

1000 runs with comm=ofi and 100 runs with comm=ugni passed, with this
change.