Add the byteIndex type and its support, make string.find() return it, and make codepoint indexing/iteration the default (#12899)
Add the byteIndex type and its support, make string.find() return it, and make codepoint indexing/iteration the default (#12899)

[reviewed by @bradcray and @mppf ]

HEADS UP:  Making string indexing default to codepoint units instead of byte units may cause some minor performance effects.

This change adds the byteIndex type and its supporting infrastructure, including arithmetic on index types to keep from having to cast back and forth just to add, subtract, or compare. This was necessary to avoid severe clutter in module code. It is still necessary to use an explicit cast to assign between index types and integers.

String indexing and iteration can now occur on a byteIndex or a byteIndex range.

This change also makes string.find() and string.rfind() return a byteIndex, and supplies quite a bit of cleanup that was required because of that.

String slicing on strided codepointIndex ranges is marked as an unimplemented feature. The string internals translate everything to byte index ranges, but a codepointIndex range is an irregular byte index range due to variable-width characters. It would require some code refactoring and the addition of a slow path to support strided codepointIndex slices, so that is a low priority feature for the future.

Codepoints are now the default units for indexing and iteration.

Tested on Mac and Linux, and on single- and multi-locale configurations.

Closes Cray/chapel-private#156