Runtime comm callbacks, chplvis datafiles support executeOn file:line.
Arrange for the line number and file number now being passed into the
executeOn functions in the runtime to be into the comm callback data and
thence into chplvis data files.  But update the chplvis tool itself only
enough to recognize that this additional data is present, and to ignore
it, because I don't feel qualified to carry it on any further (that is,
through the intake processing and into the GUI display).  At least it's
now in the data files if chplvis wants to use it, though.