Replace manual backoff spinlock for reductions with LocalSpinlock
I'm not sure why I used a backoff lock in #4532 -- It was probably just
because it was something on my mind from experiments with meteor at the
time. In this case there's no performance advantage to spinning without
yielding at least for any of our reduction tests. I also had a hard time
coming up with a test case where spinning without yielding was faster,
which is likely due to yielding getting faster as part of #4902.