Allow user to make use of known ZMQ installations when present
This changes the behavior so that CHPL_ZMQ_LOC is no longer required for
compiling multilocale libraries if ZMQ is already installed on the system and
can be found without explicit `-I` and `-L` flags on the compilation and linking
stages.  If CHPL_ZMQ_LOC is defined, use it instead of trying to find the
header and library files through traditional means.