Modify the generated `chpl_setup` function to take a numLocales argument.
Then pass that numLocales argument on to the chpl_library_init call, which will
result in spawning the server executable with the specified number of locales.
With this change, and an update to the test file to pass the number of locales,
a Python program was able to run with a multilocale library!