Optimize op= overloads for array/scalar pairs
Prior to this PR, 'A op= s' did not have a special overload, so would
get rewritten using the default implementation to 'A = A op s' which
would get rewritten as `forall (a1,a2) in zip(A, A) do a1 = a2 op s;`.
For single-dimensional arrays this was not so bad, but for
multidimensional arrays where zippered iteration is known to have
significant performance overheads, it was terrible.

This PR adds overloads for such cases, similar to the 'array = scalar'
overload that we've had for quite some time, implementing them as
`forall a in A do a op= s;`  The difference in performance is dramatic.