Avoid calling get_width() for param bool loops
By removing the case for 'bool' in get_width(), the code to handle
param for loops started hitting an internal error because it calls
get_width() on the loop's idxType even though the result isn't used
for boolean loops.  Here, I used a short-circuiting test to avoid
the call for boolean types.