Allow explicitly specifying the comm=ofi out-of-band support.
The out-of-band support is a small comm=ofi sub-interface used during
job setup and teardown, for determining the number of program instances
and the current instance's index, and for doing barrier, broadcast, and
allGather operations.  Currently there is support for PMI-, MPI-, and
sockets-based out-of-band support, although the latter is really just a
skeleton.  Up until now, out-of-band support has always been selected
automatically based on other things: PMI is used on Cray XC systems, and
MPI is used on all others (with the mpirun4ofi launcher).  Here, allow
the out-of-band support to be specified explicitly.