Add a test and future of calling nested functions within initializers
These tests were inspired by the discussion on issue #13171.  The
-correct.chpl version is kindof cool because it shows a way to push
complex initialization expressions (not that this test uses one) into
nested helper routines without exposing them to outer scopes.
However, the -incorrect.chpl version is a future that shows that we
don't check the legality of these helper routines to our detriment.