Avoid some spurious "note: because where clause evaluated to false"
When the compiler reports the reason why a function did not apply
for a particular function call, it may report "because where clause
evaluated to false". This may happen when the where-clause was added
by the compiler and not written by the user. If so, this explanation
is confusing. We prefer no explanation over a confusing explanation.
This PR makes this change.

The compiler still emits this confusing explanation when the compiler
augmented the user-written where-clause and the function did not apply
because of the compiler-added predicate evaluating to false even though
the user-written predicate evaluates to true. See #13155 and the test
discussed next:

**`test/functions/typeMethods/error-message-when-not-applicable.chpl`**

* For the purpose of #13155, it does not matter what the where-clause is.
That issue is illustrated with any where-clause that evaluates to `true`.

* The note "is passed to formal 'this' of incompatible type" in the `.good`
file is one option for the desired wording. There may be better ways
to phrase the note.