Turn off receive capabilities while creating transmit contexts.
We only transmit over our transmit contexts, but nevertheless the domain
receive attributes capabilities do indicate that receives are possible.
But the tcp provider, or perhaps the ofi_rxm utility provider used with
it, expects that when an endpoint or context which seems to be capable
of receive operations is enabled, then a CQ set up to handle those will
have been bound to it.  Failing to do that causes us to get an error
return when we enable our tx endpoint(s) and the tcp;ofi_rxm provider is
being used.

To work around this, turn off all the receive capabilities in the domain
while creating the tx endpoints and/or contexts, and then turn them back
on afterward.