Track compiler segfault when storing type of first class function directly
When the first class function is first stored into a variable and then its type
is queried, we do not encounter an issue.  However, if the type of the fcf is
stored directly, the compiler will encounter a segfault due to trying to access
a node that has already been cleaned up.