Modify the runtime for multilocale libraries
No longer link to main_mli, instead just to the regular main (since that is what
calls chpl_gen_main).  Also, change chpl_mli_smain to accept a `const char*`
instead of just a `char*` since that is the type of `c_string` under the covers
and we need to use `c_string` for chpl_gen_main to call it.

Finally, alter the runtime's handling of the connection argument to actually
store it in chpl_main_argument, since that's how we communicate argv/argc to
chpl_gen_main.