Add storage of the privacy indicator to UseStmts and update so that it gets set
This is a partial commit - these are all the updates that I think will always
be appropriate, even if the default privacy of use statements changes.  It will
not compile on its own because there are other updates that are more in flux
(either because the privacy may change or in one case due to already dropping
information on the floor that should probably be kept).