The controversial settings of a use's privacy.
We may or may not want the use of:
- the Fortran interop module
- ChapelStandard
- and the root module
to be private in the future, but that is independent of this change (and thus
should not be done yet until we verify that private uses work on their own).

Additionally, the elimination of a dead module causes the except, only, and
renaming lists to be dropped on the floor.  For now, drop privacy on the floor
(even though all of these things should be respected, since this is happening at
dead code elimination instead of pre-resolution, it doesn't actually matter)