Respond to presence of private on a use when resolving symbols (not functions)
This updates scopeResolve to notice that a use is private and not include it
when calculating uses of the module with the private use.

Also update two of the tests I wrote, since they still give the error message
from the helper file instead of just the one in the main file.