Rename moduleUseRemove to indicate it should only get called by dead code elim
This method drops information about only and except lists, renaming, and private
uses on the floor.  It isn't a big deal because the code only gets called by
dead code elimination, but we really shouldn't do it and definitely shouldn't
call it anywhere that might be pre-resolution.

Also modify the comment I wrote to be a bit better.