Remove resolveBuiltinAssignCall()
`resolveBuiltinAssignCall()` was added in #13050 (add checking for
setting non-nilable types to nil). resolveBuiltinAssignCall isn't
strictly needed but it seemed like a case the compiler should handle
directly. This transformation introduced regressions for our Intel
performance testing. We're still not sure if we're tripping over an
Intel bug or if we're generating illegal type-punning code. Until we get
to the bottom of it, remove the code that "introduced" the failures.