const checks for forall loops over ranges, domains
Resolves #13091.
Thematically related: #6142, #12944, #13268.

With this change, the index variable of a forall loop is considered
`const` when the parallel iterator has the default or `const` return intent
and the yielded things are not arrays. This is the same as what we
do for for-loops. This change applies for both "must"- and "may"-parallel
loops, i.e. loops written using the `forall` keyword and `[]` notation.
Forall loops that were using leader+follower iterators had this semantics
already prior to this change. The exclusion of arrays is discussed in #13268.

The const-checking added by this PR follows the lead of
resolveMoveForRhsSymExpr(), namely:

      if (isReferenceType(rhsType)                == false &&
          isTupleContainingAnyReferences(rhsType) == false &&
          rhsType->symbol->hasFlag(FLAG_ARRAY)    == false &&
          rhsType->symbol->hasFlag(FLAG_COPY_MUTATES) == false) {
        // ... then mark LHS constant.
        lhsSym->addFlag(FLAG_CONST);
      }

except no checking for isReferenceType(rhsType) and instead checking
the iterator's return intent, which is stored in its IteratorInfo.

While there, add another future test for #6142.

Trivial, not reviewed.