Fix support for llvm+gasnet on Crays
For llvm configurations, we build the runtime with CC=clang-included in
order to have a runtime that's compatible with llvm backend. However,
the cross-configure scripts we use to build gasnet on Crays force set
CC=cc so the PrgEnv wrappers are always used. Here we make sure that we
use PrgEnv-gnu because we know it should be ABI and compiler option
compatible.

Note that we're not just forcing gasnet to use clang-included because we
haven't been able to get clang-included to compile MPI programs so we
lose MPI compatibility and couldn't build the gasnet-mpi config.