Allow in-intents of POD record types in follower iterators
Resolves #13169.
Other references: #13293, #13299.

Background: our _toFollower function invokes the follower iterator
passing to it the formals of the serial iterator. This gets the job
done, however it is a poor representation because it uses those
formals outside of their scope. lowerIterators removes these
references by replacing them with fields of the corresponding _iteratorClass.
This replacement was missing the formal-temp for the in-intent formal
in the case it was created.

This change turns off formal temps in this particular scenario,
resolving this issue.