Permit 'string' type to be passed to externs, just not 'string' values
We've historically prohibited passing Chapel string values to extern
procedures, but the check that did this was also preventing the Chapel
string type to extern procedures, which I think is an oversight /
overkill (and something Louis was wanting to do in #13264, which I
think is reasonable).

While here, note that I think in the future that we may want to start
permitting Chapel strings to be passed to external procedures, and to
have them unwrap the string as a dual to issue #12892.  But that's
a separate effort.