In comm=ofi, make the PMI out-of-band support more principled.
The comm=ofi, the PMI-based out-of-band support was really only
appropriate for Cray XC systems.  It mixed use of the PMI and PMI2
interfaces, and depended upon PMI_Barrier(), PMI_Allgather(), and
PMI_Bcast() functions that are available only in the PMI_*() interface
and/or only as Cray extensions.  Here, to aid portability, change to
using only the PMI2_*() interface and only base functions, nothing that
is a Cray extension.  Notably, this should allow us to launch comm=ofi
programs using the slurm-srun launcher on systems where the slurm PMIx
plugin is being used.