Type Constructor Implementation No Longer Uses FnSymbol
This commit updates the compiler to no longer use a FnSymbol to represent and
resolve a type constructor call. Instead, the compiler represents such calls
with a CallExpr whose 'baseExpr' points to the TypeSymbol we wish to
instantiate and resolve. This CallExpr is passed to methods on AggregateType
that then take over the process of evaluating arguments and creating an
instantiation.

The 'generateType' methods on AggregateType have been updated to take over the
process of type-checking the fields' type-expressions and default-expressions.
Instantiating a type now causes the fields' type/default-expressions to be
normalized, resolved, and checked to ensure that the result is compatible with
the given argument for a field.

In order to support clean error messages we now compute a string during
scopeResolve that looks like the old FnSymbol signature. This string, along
with CallInfo::toString, is used to mimic the error style for typical
resolution failures.

The FnSymbol representation had features that needed to be reproduced in this
new representation:
- AggregateType has a new field, 'resolveStatus', used to prevent recursive
  resolution.
- AggregateType has a new field, 'genericFields', which keeps track of the
  generic fields. This is used in places that relied on the FnSymbol's formals.

Miscellaneous changes were required to better handle this new representation:
- Update CallExpr::qualType to better handle new type constructor call format
- Mark LoopExpr functions as generic manually as they may be generated during
  field normalization, but after the point where the compiler would normally
  do that for us.
- Change 'update_symbols' to not replace the baseExpr of type constructor calls
  (except in some special cases, like tuples)
- Add 'dtRef' for easier creation of type-constructor calls
- Update TypeSymbol::copyInner to preserve parentSymbol/Expr links for fields
- Using TypeSymbol::instantiationPoint for fields in getInstantiationPoint

Vestiges of the old representation, like flags or '_type_construct' strings,
have been removed along with code paths dedicated to the FnSymbol
representation.