Remove chpl_emptyLocaleSpace and chpl_emptyLocales from ChapelDistribution
These constants were also defined in the flat and apu LocaleModels, and having
them in ChapelDistribution caused a circular dependency resolution error when
trying to make the use of ChapelStandard private.

I need to define them in the numa and knl LocaleModels but this is to check that
removing them doesn't cause problems for flat.