Remove prototype HDFS code
Most of the functionality is in the new QIO plugin but there remains code
especially relating to getting affinity that is not moved over. However
the affinity code depends on the use case of Chapel jobs being launched
on the same nodes that are storing the HDFS data which likely has other
challenges.