Move LocaleModel.writeThis to ChapelIO
Prior to this change, each LocaleModel would individually define identical
writeThis functions.  When I tried to make the use of ChapelStandard private,
we started encountering circular dependencies because of this, due to not
explicitly using IO and the IO module relying on things the LocaleModel would
define.  This change moves the writeThis function to ChapelIO to avoid this
circular dependency.