Use comm=ofi MPI-based out-of-band support on Cray Shasta systems.
PMI seems not yet ready for prime time.  For now use MPI-based
out-of-band support.  Note that we don't want to do this forever,
because (I suspect) a side effect will be that we'll end up with one or
more MPI progress threads running.  But at least programs will link and
start dependably.