Change how the use of ISO_Fortran_binding is inserted for --library-fortran
Prior to this change, when compiling with --library-fortran we were inserting a
use of ISO_Fortran_binding into every standard module and every top level user
code module.  This was causing circular dependencies when I made the use of
ChapelStandard private, because CommDiagnostics is the first standard module
used and it turns up very early in ChapelStandard's chain, before the modules
ISO_Fortran_binding would rely on have been initialized. Instead, let's only
insert a use of ISO_Fortran_binding as the last statement in ChapelStandard.