At Michael's suggestion, a slight shuffling of the uses
Instead of using these modules from standard modules, make them available via
ChapelStandard - they already were available, but it is not the perogative of
this change to alter that and making them private uses in the standard modules
would do so.  We can then discuss whether SysBasic or SysError should be
included by default.