Print modUseList in AST logs, change it to std::vector
Including modUseList in the AST logs helps to understand
problems with code using it. This commit adds it to the
--log output. However, in doing so, I was running into a
problem with using a const Vec and so I migrated modUseList
to a std::vector. That was mostly straightforward but
there is no std::vector equivalent to `index` so I added
a static helper function to ModuleSymbol.cpp for that.