Resolve errors from collected methods going away
The baseAST code for removing a FnSymbol will set the
relevant value in methods or dispatchChildren to NULL.
It might be reasonable for the inter-pass cleanup/GC
to pack these lists also but for now the holes remain.
This commit adapts code to check against NULL and ignore
such values when traversing these lists.