Switch "generic actual" internal -> user error
Resolves #11410.
Illustrates that #11846 is resolved.

Make the error "the type of the actual argument 'XXX' is generic"
a user error instead of a compiler internal error.

Because this error can arise when compiling a program exhibiting this issue.
Which is illustrated by the tests affected by this change.

The sister error "use of 'XXX' before encountering its definition"
was converted from compiler internal error to user error in 0de048e68d.

The following tests, while hitting this error, are for different topics,
as indicated by their .future files:

    test/classes/ferguson/class-alias-generic-arg.chpl
    test/classes/generic/nestedGenericTypes.chpl
    test/functions/vass/declaredGenericReturnTuple.chpl
    test/types/enum/scope-param-offset-type.chpl
    test/users/thom/yieldGeneric.chpl