Remove the future test dmap-var-decl.chpl
This test was added in 2011 and had somewhat of a rocky history.
Most recently, mid-2018 through 2019-04-07 or so it was tested only on darwin
and generated `dmap-var-decl.chpl:2: error: attempt to dereference nil` .
During a recess of darwin testing in April/May, the following line:

```chpl
var MyBlockDist: dmap(Block(rank=2));
```

started causing compilation errors. First off, the type within `dmap()`
must be nilable, because `MyBlockDist` does not provide any non-nil instance.

With that change, i.e. `dmap(Block(rank=2)?)`, compiling with
--no-legacy-nilable-classes and tweaking the modules, ex.

```diff
-  proc chpl__buildDistType(type t) type where isSubtype(t, BaseDist) {
+  proc chpl__buildDistType(type t) type where isSubtype(t!, BaseDist) {

// in record _distribution:
-    proc init(value) {
+    proc init(_instance) {
```

makes it clear that our modules currently do not support `dmap`
with a nilable type. Which is what the code above is doing.
We will adjust the modules to support this when/if there is user demand.
Meanwhile I am removing this test.

Line 3 in this test exercises a non-nilable version of the above:

```chpl
var MyBlockDist2: dmap(Block(rank=2)) = new dmap(new Block({1..5,1..6}));
```

This pattern is covered in other tests, for example:

    test/distributions/deitz/test_distribution_syntax2.chpl

and also as a field in a class:

    test/classes/initializers/phase1/domainMaps/init-dmap-dom-arr.chpl