Use a blocking progress thread for gasnet-infiniband
Avoid interference from an active polling/progress thread by using a
blocking progress thread. Here we use `GASNET_RCV_THREAD` to tell gasnet
to create another thread that will block while waiting for incoming AMs.
This eliminates contention from the progress thread when there are no
incoming AMs.

Note that this thread only services network AMs, and would not progress
a shared-memory channel. This isn't a problem for us since we currently
run with a process per node and disable shared-memory support for IBV
with `--disable-pshm`.

This slightly increases the latency of AMs for gasnet-ibv, but
eliminating interference has non-trivial improvements for most
multi-locale benchmarks. The performance trade-offs are similar to
what we saw when we made a similar change for ugni in #9068.