Add error overload for attempts to provide a type as a domain expression
This adds an error overload for the compiler's internal
chpl__ensureDomainExpr() routine that accepts a type to generate a
useful error message in the event that a type expression is provided
rather than a domain value or range list as currently expected.

With more work, additional error overloads could be added to head off
other mistakes, but I'm only handling this one for now since users
have run into it and it relates to an idea proposed in issue #5355.