Stop pinning the progress thread to the last CPU
In 2425 we decided to pin the progress/polling thread to the last
available CPU with the thought that it would move it out of the way and
let us more fully utilize cores.

Since then (especially with spectre/meltdown patches) we have seen
significant interference from the progress thread and pinning it just
puts all the interference on one core, which makes tasks running on that
core significantly slower. Stop pinning, and let the kernel decide where
to run in hopes of spreading the pain.

We're also moving towards using blocking progress threads since there is
still non-trivial overhead from active polling even with this change. We
already do this for ugni, and are workings towards that for gasnet
configurations that support it.

I expect non-trivial performance improvements for compute heavy
benchmarks under gasnet and no changes for ugni.