Move main process sleeping code into common shim
For ugni and gasnet we put the main process to sleep on a condition
variable in order to free up `chpl_comm_barrier()` for use in module
code and to avoid interference from the main process. This code is
duplicated across comm layers and needs to be used in ofi soon, so just
move most of the code into helper routines.