Specify default transmit and receive completion types.
We were getting completions for transmits, because we were waiting to
see them in CQ events and surely would have hung had they not arrived.
But it's not clear what kind we were getting - the man pages don't say
what the defaults are and it may be provider-specific.  What is clear is
that in certain cases we were getting them earlier than when the target
side had the data, which would be FI_DELIVERY_COMPLETE.  This was the
cause of the miscommunications in hpl and others with the 'tcp;ofi_rxm'
provider.  Here, explicitly specify delivery completions for transmits.

While we're at it, also explicitly specify the need for multi-receive
completions on receives.  It's likely this is redundant since we also
specify this in the main capabilities, but there's no harm is making
sure.