Split native AMO logic to make the next change prettier.
We do slightly different logic for native compare, non-compare fetching,
and non-fetching AMOs.  Here, rearrange so that we do the differing
(compare, fetch, non-fetch) part up to and including the fi_atomic()
calls, then common code for the completion wait, and then the differing
(fetch, non-fetch) part for the result handling.  The completion waiting
is going to get more complex in the next commit, and I don't want to
have to repeat it 3 times as I would with the current arrangement.