Revamp transaction ordering.
This is a major rework of transaction ordering, which should make things
correct while also allowing for future optimization.  We still wait for
each transaction to complete before starting the next.  But the waiting
now uses 'done' flags or task-private counters instead of counters in
the transmit context descriptors, which couldn't work properly when
tasks could switch transmit contexts (as with tasks=fifo).  'Done' flags
are used for all transactions (GETs, PUTs, AMOs, AMs) now, but there is
(untested) logic to use task-private counters for unordered transactions
in the future.

There is one exception to the above.  The AM handler's transmit context
has a completion counter rather than a completion queue, in order to
reduce overhead.  So here we still use a counter in the transmit context
descriptor for transaction waiting.  This is safe because the AM handler
has a private transmit context descriptor and thus is not subject to
context switching.  That said, maximizing comm/compute overlap in the AM
handler in the future will probably require changing to a completion
queue in order to track individual transactions.