Fix a problem involving error handling and default arguments
When moving default argument values for a function ahead of the call site
so they can be passed directly, the compiler was moving a throwing call from
inside a 'try!' or 'try' expression to just before it.

When creating the default argument temp, check if the involved call throws.
If so, then if the old location was inside a 'try' or 'try!' create a new
'try' or 'try!' to put the new call into.

Add a test that previously hit compiler errors about calling throwing
functions outside of a 'try'/'try!' for a compiler-generated function.