Remove old ugni chpl_comm_get_nb declaration
Originally chpl_comm_get_nb was only implemented in ugni, but it was
made a part of the official comm API years ago in bcf2c83d9b. Back then
ugni was still private, and it looks like we missed removing a decl and
the old style of waiting for a NB get. Remove the old decls from ugni
(and ofi since that file was originally a clone of the ugni one.)