Forall intents to operate on fields of 'this'
As suggested here:

    https://github.com/chapel-lang/chapel/issues/13297#issuecomment-505208697

specifically, "we don't think about the intent for `this` when considering
`data` within the loop, we just think about the intent for `data`
(as an array, its passed by reference)".

When a forall loop operates on the fields of the `this` formal, handle these
fields as if they are regular variables passed by the default intent.
This uses `ref` temporaries. This applies only when `this` is a record.
For example, convert:

    forall ... {
      this.data = i;
    }

to

    ref data_ref = this.data;
    forall ... {  // 'data_ref' is passed by the default intent
      data_ref = i;
    }

Cf. before this change the code above would be converted to:

    forall ... with (const ref this) {
      this.data = i;  // 'this' is a shadow variable.
    }

At present this is not implemented when `this` is of an array type,
even though the compiler considers it a "record".