Add setSubscribe and setUnsubscribe implementations, to replace setsockopt
Also extends the warning on setsockopt's documentation to link to all three
alternative versions instead of just setLinger.

I added a version for POD types and for strings, to match the implementations
that were provided for setsockopt - the description in the ZMQ API says that it
accepts "binary data" so either of those are appropriate (and no more or less
than we already provided).