Privately use TomlReader as the module is marked "no doc"
TomlParser is not marked "no doc" and has some symbols that are documented and
visible, so it does not get the same treatment.  This seems to be safe to do
based on the contents of the library/packages/TOML/test directory, but I will
need a full test run to be certain.