Make use of Regexp private to the helper modules
Regexp is used in the implementation of things that are hidden from outside
the package module, so does not need to be exposed outside of the context in
which it is used.

The symbols from the DateTime module are exposed as part of the user-facing
interface in the TomlParser module, so should remain publicly used for
convenience.