Fix a failure to report an error when a default argument value throws
When a function has a default argument that throws but the function is not
marked 'throws', it should be reported as a compile time error. Default
argument values are pulled out into wrapper functions that are called before
the original function. When creating this wrapper, mark it with a new flag
FLAG_DEFAULT_ACTUAL_FUNCTION. When issuing errors about throwing from
functions not marked with 'throws', include functions with this flag even
though they also have FLAG_COMPILER_GENERATED.

Added a test that previously failed to issue the desired error, but now does.