Add a note to guard against a concern of Michael's for the ZMQ module
Michael wants to be certain that if we fail to provide support for a particular
ZMQ option, the user has a way to work around it until we do support it.  With
setsockopt going away, the current strategy relies on the Socket classRef field
not being private.  Leave a note on that field so we don't forget about that
benefit when the day comes where we can make it private.