Make the use of the Sys module private to Path
Path uses sys_getenv in its implementation of some functions.  There is no need
to expose this use further.

I chose to leave the use of SysError public due to the functions in this module
being able to throw errors from it.  We discussed this in relation to the ZMQ
module on #13495.

I chose to leave the use of the FileSystem module as is - it is within a
function, and there are no other uses of it in the rest of the module.