Privately use the MemConsistency module in ChapelSyncvar
This module uses the chpl_rmem_* methods as part of its implementation, and does
not expose its use of this module in other ways.  It does not need to expose its
use of this module.

I left the other uses in this module alone - AlignedTSupport and
SyncVarRuntimeSupport are both private modules, so using them does not expose
their symbols further.  The uses in SyncVarRuntimeSupport are also not really
public due to the module itself being private - they may impact the visibility
of symbols in ChapelSyncvar, but ChapelSyncvar already uses both of these
modules so is unlikely to be impacted either way.