Adjust cullOverReferences for proper const?-ref
With this commit, in this code:

    proc MyRecord.myProc() {
      forall .... {
        ... use( this.myArrayField ) ...;
      }
    }

the intent for `this` in `myProc` is 'ref' if use() is a write
and 'const ref' if use() is a read.

DETAILS

With the previous commit 43cd18b766, the compiler rewrites the above into:

    proc MyRecord.myProc() {
      ref myArrayField_ref = this.myArrayField
      forall .... with (ref myArrayField_svar(outerVar = myArrayField_ref)) {
        ... use( myArrayField_svar ) ...;
      }
    }

One part of this change to cullOverReferences is to build the dependence
from myArrayField_svar to myArrayField_ref upon seeing the relationship
between these two as the shadow variable and the outer variable
in the ForallStmt.

The lack of this dependence was not an issue earlier because the
compiler replaces references to a ref-intent shadow variable with references
directly to its outer variable and discards the shadow variable, see
'prune' in resolveShadowVarTypeIntent(). For this change, I chose not to do so
in order to start exercising ref-intent shadow variables in the rest
of the compiler. Although I am not completely sure that **all**
ref-intent shadow variables were pruned before.

While "ref" in `myArrayField_ref` did not side-track cullOverReferences into
forcing the ref-intent on the `this` formal, "ref" in myArrayField_svar did.
Below I list the source and the AST at start of cullOverReferences
for two cases: (A) an implicit shadow variable for an array field
and (B) an array formal with const?-ref intent.

symExprIsSet(), invoked in (A) on SymExpr 925541, used to return true
just because `xx_svar` has FLAG_REF_VAR. Which it should, as it has
a ref intent. When invoked in (B) on SymExpr 925548 in an otherwise-identical
situation, returns false because `arg` is a formal and so does not have
FLAG_REF_VAR even though it also has a ref intent.
I apply FLAG_DEFAULT_INTENT_IS_REF_MAYBE_CONST to obtain the desired
behavior in symExprIsSet() so that `this` formal is not considered modified
in this case.

-----------------------------------------------------------------------------

(A) an implicit shadow variable for an array field `xx`

const RNG = 1..4;

record RR {
  var xx: [RNG] int;
}

proc RR.method() {
  forall idx in RNG do
    writeln(xx[idx]);
}

var ff: RR;
proc main {
  ff.method();
  writeln(ff);
}

181023  fn method[181023]:void[4]
181027  def ref arg intent-'const? ref' this[181025]:_ref(RR)[916895]
181024  {
924436    def ref xx_ref[924435]:_ref(_array())[923730]
924442    move( xx_ref[924435] call( fn xx[613954] arg intent-'const? ref' this[181025] ) )

181040    forall
181042      def const idx[181041]:int(64)[13]
          in
924423      call( fn these[754299] RNG[181002] )
          with
924449      def ref xx_svar[924446]:_array()[923704]
924445        xx_ref[924435]
          do
181039      {
535788        def ref call_tmp[535787]:_ref(int(64))[644823]
535790        move( call_tmp[535787] ...) -- see details:

	      (535790 CallExpr move
		(535791 SymExpr 'ref call_tmp[535787]:_ref(int(64))[644823]')
		(925583 ContextCallExpr
		  (925539 CallExpr
		    (925581 SymExpr 'fn this[924951]:int(64)[13]')
		    (925541 SymExpr 'ref xx_svar[924446]:_array()[923704]')
		    (925542 SymExpr 'const idx[181041]:int(64)[13]'))
		  (925560 CallExpr
		    (925582 SymExpr 'fn this[925158]:_ref(int(64))[644823]')
		    (925562 SymExpr 'ref xx_svar[924446]:_array()[923704]')
		    (925563 SymExpr 'const idx[181041]:int(64)[13]'))
		  (181036 CallExpr
		    (925580 SymExpr 'fn this[924853]:_ref(int(64))[644823]')
		    (924450 SymExpr 'ref xx_svar[924446]:_array()[923704]')
		    (285214 SymExpr 'const idx[181041]:int(64)[13]'))))

929400        def val coerce_tmp[929398]:int(64)[13]
929402        move( coerce_tmp[929398] deref( call_tmp[535787] ) )
181038        call( fn writeln[929380] coerce_tmp[929398] )
181039      }

378387    return( _void[45] )
181024  }

-----------------------------------------------------------------------------

(B) an array formal `arg` with const?-ref intent

const RNG = 1..4;

type RR = [RNG] int;

proc method(arg: RR) {
  forall idx in RNG do
    writeln(arg[idx]);
}

var ff: RR;
proc main {
  method(ff);
  writeln(ff);
}

181023  fn method[181023]:void[4]
181022  def ref arg intent-'const? ref' arg[181020]:_array()[924147]
181024  {

181035    forall
181037      def const idx[181036]:int(64)[13]
          in
924451      call( fn these[754626] RNG[181002] )
          with() do
181034      {
536483        def ref call_tmp[536482]:_ref(int(64))[645150]
536485        move( call_tmp[536482] ...)

	      (536485 CallExpr move
		(536486 SymExpr 'ref call_tmp[536482]:_ref(int(64))[645150]')
		(925590 ContextCallExpr
		  (925546 CallExpr
		    (925588 SymExpr 'fn this[924958]:int(64)[13]')
		    (925548 SymExpr 'ref arg arg[181020]:_array()[924147]')
		    (925549 SymExpr 'const idx[181036]:int(64)[13]'))
		  (925567 CallExpr
		    (925589 SymExpr 'fn this[925165]:_ref(int(64))[645150]')
		    (925569 SymExpr 'ref arg arg[181020]:_array()[924147]')
		    (925570 SymExpr 'const idx[181036]:int(64)[13]'))
		  (181031 CallExpr
		    (925587 SymExpr 'fn this[924860]:_ref(int(64))[645150]')
		    (285198 SymExpr 'ref arg arg[181020]:_array()[924147]')
		    (285199 SymExpr 'const idx[181036]:int(64)[13]'))))

929407        def val coerce_tmp[929405]:int(64)[13]
929409        move( coerce_tmp[929405] deref( call_tmp[536482] ) )
181033        call( fn writeln[929387] coerce_tmp[929405] )
181034      }

378372    return( _void[45] )
181024  }

-----------------------------------------------------------------------------